+++ doc/CHANGES.rst
@@ -25,9 +25,12 @@ CHANGELOG
       class MyModelInOtherSchema:
           __db_schema__ = 'my_schema'
 
-* Now the Email, Selection, Country fields can be crypted
+
+* Now the Email, Selection, Country, Json, Decimal, Interval, LargeBinary 
+  columns can be crypted
 * Removed compatibility with **Python 3.4**, because this version is
   deprecated
+* Added *TimeStamp* column
 
 0.22.5 (2019-06-24)
 -------------------